pyproject.toml
[project]
name = "reporadar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.5",
    "gitpython>=3.1.45",
]
